feat: spliting the main game function to an Archive at src/Util
